+++ dzgui/controllers/mc.py
@@ -1,4 +1,5 @@
 import logging
+import re
 import shutil
 import threading
 import textwrap
@@ -43,6 +44,8 @@ from dzgui.util import cooldown, strings
 from dzgui.util._json import read_json, write_json
 from dzgui.util.open_links import open_workshop_page
 from dzgui.util.format import format_mods
+from dzgui.util.redact import redact_log
+
 from dzgui.views.dialogs.filepicker import FilePicker
 from dzgui.views.dialogs.generic import GenericDialog
 
@@ -437,7 +440,8 @@ class Controller:
         with open(log, "r") as f:
             lines = [line.split(strings.delimiter) for line in f.read().splitlines()]
             for record in lines:
-                store.append(record)
+                clean = redact_log(record)
+                store.append(clean)
         self.open_page(NotebookPage.LOG)
 
     def select_colorized(self) -> None: